// src/app/api/api.ts
export const API_BASE_URL = new InjectionToken<string>('API_BASE_URL');
// ...
export class OptionVehicleAPI {
    private http: HttpClient;
    private baseUrl: string;
    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;

    constructor(@Inject(HttpClient) http: HttpClient, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
        this.http = http;
        this.baseUrl = baseUrl ? baseUrl : "http://localhost:63468";
    }

    /**
     * @return OK
     */
    getOptionVehicleList(): Observable<OptionVehicleDTO[]> {
        let url_ = this.baseUrl + "/OptionVehicle/List";
        url_ = url_.replace(/[?&]$/, "");

        let options_ : any = {
            observe: "response",
            responseType: "blob",
            headers: new HttpHeaders({
                "Accept": "application/json"
            })
        };

        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
            return this.processGetOptionVehicleList(response_);
        })).pipe(_observableCatch((response_: any) => {
            if (response_ instanceof HttpResponseBase) {
                try {
                    return this.processGetOptionVehicleList(<any>response_);
                } catch (e) {
                    return <Observable<OptionVehicleDTO[]>><any>_observableThrow(e);
                }
            } else
                return <Observable<OptionVehicleDTO[]>><any>_observableThrow(response_);
        }));
    }

    protected processGetOptionVehicleList(response: HttpResponseBase): Observable<OptionVehicleDTO[]> {
        const status = response.status;
        const responseBlob = 
            response instanceof HttpResponse ? response.body : 
            (<any>response).error instanceof Blob ? (<any>response).error : undefined;

        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }};
        if (status === 200) {
            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
            let result200: any = null;
            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
            if (Array.isArray(resultData200)) {
                result200 = [] as any;
                for (let item of resultData200)
                    result200!.push(OptionVehicleDTO.fromJS(item));
            }
            return _observableOf(result200);
            }));
        } else if (status !== 200 && status !== 204) {
            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            }));
        }
        return _observableOf<OptionVehicleDTO[]>(<any>null);
    }
}
// ...
export class VehicleDTO implements IVehicleDTO {
    id?: string | undefined;
    year?: number | undefined;
    mileage?: number | undefined;
    modelId?: string | undefined;
    plateNumber?: string | undefined;
    chassisNumber?: string | undefined;
    opinionAverage?: number | undefined;
    optionNames?: string[] | undefined;
    optionVehicleVehicles?: OptionVehicleVehicleDTO[] | undefined;
    model?: ModelDTO | undefined;
    maintenances?: MaintenanceDTO[] | undefined;

    constructor(data?: IVehicleDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.year = data["Year"];
            this.mileage = data["Mileage"];
            this.modelId = data["ModelId"];
            this.plateNumber = data["PlateNumber"];
            this.chassisNumber = data["ChassisNumber"];
            this.opinionAverage = data["OpinionAverage"];
            if (Array.isArray(data["OptionNames"])) {
                this.optionNames = [] as any;
                for (let item of data["OptionNames"])
                    this.optionNames!.push(item);
            }
            if (Array.isArray(data["OptionVehicleVehicles"])) {
                this.optionVehicleVehicles = [] as any;
                for (let item of data["OptionVehicleVehicles"])
                    this.optionVehicleVehicles!.push(OptionVehicleVehicleDTO.fromJS(item));
            }
            this.model = data["Model"] ? ModelDTO.fromJS(data["Model"]) : <any>undefined;
            if (Array.isArray(data["Maintenances"])) {
                this.maintenances = [] as any;
                for (let item of data["Maintenances"])
                    this.maintenances!.push(MaintenanceDTO.fromJS(item));
            }
        }
    }

    static fromJS(data: any): VehicleDTO {
        data = typeof data === 'object' ? data : {};
        let result = new VehicleDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Year"] = this.year;
        data["Mileage"] = this.mileage;
        data["ModelId"] = this.modelId;
        data["PlateNumber"] = this.plateNumber;
        data["ChassisNumber"] = this.chassisNumber;
        data["OpinionAverage"] = this.opinionAverage;
        if (Array.isArray(this.optionNames)) {
            data["OptionNames"] = [];
            for (let item of this.optionNames)
                data["OptionNames"].push(item);
        }
        if (Array.isArray(this.optionVehicleVehicles)) {
            data["OptionVehicleVehicles"] = [];
            for (let item of this.optionVehicleVehicles)
                data["OptionVehicleVehicles"].push(item.toJSON());
        }
        data["Model"] = this.model ? this.model.toJSON() : <any>undefined;
        if (Array.isArray(this.maintenances)) {
            data["Maintenances"] = [];
            for (let item of this.maintenances)
                data["Maintenances"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IVehicleDTO {
    id?: string | undefined;
    year?: number | undefined;
    mileage?: number | undefined;
    modelId?: string | undefined;
    plateNumber?: string | undefined;
    chassisNumber?: string | undefined;
    opinionAverage?: number | undefined;
    optionNames?: string[] | undefined;
    optionVehicleVehicles?: OptionVehicleVehicleDTO[] | undefined;
    model?: ModelDTO | undefined;
    maintenances?: MaintenanceDTO[] | undefined;
}
// ...
export class OptionVehicleVehicleDTO implements IOptionVehicleVehicleDTO {
    id?: string | undefined;
    vehicleId?: string | undefined;
    optionVehicleId?: string | undefined;
    vehicle?: VehicleDTO | undefined;
    optionVehicle?: OptionVehicleDTO | undefined;

    constructor(data?: IOptionVehicleVehicleDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.vehicleId = data["VehicleId"];
            this.optionVehicleId = data["OptionVehicleId"];
            this.vehicle = data["Vehicle"] ? VehicleDTO.fromJS(data["Vehicle"]) : <any>undefined;
            this.optionVehicle = data["OptionVehicle"] ? OptionVehicleDTO.fromJS(data["OptionVehicle"]) : <any>undefined;
        }
    }

    static fromJS(data: any): OptionVehicleVehicleDTO {
        data = typeof data === 'object' ? data : {};
        let result = new OptionVehicleVehicleDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["VehicleId"] = this.vehicleId;
        data["OptionVehicleId"] = this.optionVehicleId;
        data["Vehicle"] = this.vehicle ? this.vehicle.toJSON() : <any>undefined;
        data["OptionVehicle"] = this.optionVehicle ? this.optionVehicle.toJSON() : <any>undefined;
        return data; 
    }
}

export interface IOptionVehicleVehicleDTO {
    id?: string | undefined;
    vehicleId?: string | undefined;
    optionVehicleId?: string | undefined;
    vehicle?: VehicleDTO | undefined;
    optionVehicle?: OptionVehicleDTO | undefined;
}

export class ModelDTO implements IModelDTO {
    id?: string | undefined;
    name?: string | undefined;
    displacement?: number | undefined;
    weight?: number | undefined;
    price?: number | undefined;
    increasePrice?: number | undefined;
    distanceMax?: number | undefined;
    pictureLink?: string | undefined;
    brandId?: string | undefined;
    brand?: BrandDTO | undefined;

    constructor(data?: IModelDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.name = data["Name"];
            this.displacement = data["Displacement"];
            this.weight = data["Weight"];
            this.price = data["Price"];
            this.increasePrice = data["IncreasePrice"];
            this.distanceMax = data["DistanceMax"];
            this.pictureLink = data["PictureLink"];
            this.brandId = data["BrandId"];
            this.brand = data["Brand"] ? BrandDTO.fromJS(data["Brand"]) : <any>undefined;
        }
    }

    static fromJS(data: any): ModelDTO {
        data = typeof data === 'object' ? data : {};
        let result = new ModelDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Name"] = this.name;
        data["Displacement"] = this.displacement;
        data["Weight"] = this.weight;
        data["Price"] = this.price;
        data["IncreasePrice"] = this.increasePrice;
        data["DistanceMax"] = this.distanceMax;
        data["PictureLink"] = this.pictureLink;
        data["BrandId"] = this.brandId;
        data["Brand"] = this.brand ? this.brand.toJSON() : <any>undefined;
        return data; 
    }
}
// ...
export class OptionVehicleDTO implements IOptionVehicleDTO {
    id?: string | undefined;
    name?: string | undefined;
    optionVehicleVehicles?: OptionVehicleVehicleDTO[] | undefined;

    constructor(data?: IOptionVehicleDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.name = data["Name"];
            if (Array.isArray(data["OptionVehicleVehicles"])) {
                this.optionVehicleVehicles = [] as any;
                for (let item of data["OptionVehicleVehicles"])
                    this.optionVehicleVehicles!.push(OptionVehicleVehicleDTO.fromJS(item));
            }
        }
    }

    static fromJS(data: any): OptionVehicleDTO {
        data = typeof data === 'object' ? data : {};
        let result = new OptionVehicleDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Name"] = this.name;
        if (Array.isArray(this.optionVehicleVehicles)) {
            data["OptionVehicleVehicles"] = [];
            for (let item of this.optionVehicleVehicles)
                data["OptionVehicleVehicles"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IOptionVehicleDTO {
    id?: string | undefined;
    name?: string | undefined;
    optionVehicleVehicles?: OptionVehicleVehicleDTO[] | undefined;
}